fix operator prompt scroll style
